Update django-tinymce to use the latest github commit
The current django-tinymce in PyPI uses a vulnerable version of TinyMCE.
The Github repo has an updated version, but it has not been released to
PyPI yet. This commit makes the application use the github version. The
dependencies should be updated once the next django-tinymce version is
uploaded on PyPI.